+++ example/pirate/intelligence/configs.py
@@ -14,7 +14,7 @@ OLLAMA_LLAMA_3_2 = OllamaLlmConfig(
 OLLAMA_LLAMA_3_1 = OllamaLlmConfig(
     host=os.getenv("OLLAMA_HOST"),
     port=int(os.getenv("OLLAMA_PORT", 11434)),
-    model='llama3.1:8b-instruct-q8_0',
+    model='llama3.1:8b-instruct-q4_K_M',
     temperature=1.0,
     prompt_retry_count=3,
 )